requirements.txt
python-dotenv
mysql-connector-python
numpy
matplotlib
